refactoring according to MKasaei00's comments
